feat: multiple schema tests

+++ services/nihongo-dico-entries-service/src/main/resources/db/migration/public/V1.0__ddl.sql
@@ -2,7 +2,7 @@ CREATE TABLE about
 (
     about_id    SERIAL PRIMARY KEY,
     jmdict_date TEXT    NOT NULL,
-    nb_entries  NUMERIC NOT NULL,
+    nb_entries  INTEGER NOT NULL,
     languages   json    NOT NULL
 );
